+++ src/69while.js
@@ -1,0 +1,74 @@
+/*
+//
+// CREATE VIEW for Alasql.js
+// Date: 03.11.2014
+// (c) 2014, Andrey Gershun
+//
+*/
+
+yy.While = function (params) { return yy.extend(this, params); }
+yy.While.prototype.toString = function() {
+	var s = 'WHILE ';
+	s += this.expression.toString();
+	s += ' '+this.loopstat.toString();
+	return s;
+};
+
+yy.While.prototype.execute = function (databaseid,params,cb,scope) {
+	var self = this;
+	var res;
+//	console.log(this.expression.toJavaScript());
+	var fn = new Function('params,alasql','return '+this.expression.toJavaScript());
+	var res = fn(params,alasql);
+	if(res) {
+		res = this.loopstat.execute(databaseid,params,cb,scope);
+	} else {
+		if(cb) res = cb(res);
+	}
+	return res;
+};
+
+yy.Break = function (params) { return yy.extend(this, params); }
+yy.Break.prototype.toString = function() {
+	var s = K('BREAK');
+	return s;
+};
+
+yy.Break.prototype.execute = function (databaseid,params,cb,scope) {
+	var res = 1;
+	return res;
+};
+
+yy.Continue = function (params) { return yy.extend(this, params); }
+yy.Continue.prototype.toString = function() {
+	var s = K('CONTINUE');
+	return s;
+};
+
+yy.Continue.prototype.execute = function (databaseid,params,cb,scope) {
+	var res = 1;
+	return res;
+};
+
+yy.BeginEnd = function (params) { return yy.extend(this, params); }
+yy.BeginEnd.prototype.toString = function() {
+	var s = K('BEGIN')+' '+this.statements.toString()+' '+K('END');
+	return s;
+};
+
+yy.BeginEnd.prototype.execute = function (databaseid,params,cb,scope) {
+	var self = this;
+	var res = [];
+
+	var idx = 0;
+	runone();
+	function runone() {
+		self.statements[idx].execute(databaseid,params,function(data){
+			res.push(data);
+			idx++;
+			if(idx<self.statements.length) runone();
+			if(cb) res = cb(res);
+		});
+	}
+	return res;
+};