+++ CHANGELOG.md
@@ -43,7 +43,7 @@
 * alasql.work(alasql.js, [plugins...], callback)
 * @localvariable
 * SET @localvariable = expression
-
+* WHILE, CONTINUE, BREAK, BEGIN END syntax
 
 ### 0.0.35 (14.12.2014 - 22.12.2014)